+++ packages/fs/src/hyperbook.ts
@@ -84,13 +84,25 @@ export const getNavigationForFile = async (
     i = pageList.findIndex((p) => p.href === currentFile.path.href);
 
     if (current !== null && current.next !== undefined) {
-      next = pageList.find((p) => p.href === current?.next) || null;
+      next =
+        pageList.find((p) => {
+          if (current?.next?.startsWith("/@/")) {
+            return p.permaid === current.next.split("/@/")[1].trim();
+          }
+          return p.href === current?.next;
+        }) || null;
     } else {
       next = pageList[i + 1] || null;
     }
 
     if (current !== null && current.prev !== undefined) {
-      previous = pageList.find((p) => p.href === current?.prev) || null;
+      previous =
+        pageList.find((p) => {
+          if (current?.prev?.startsWith("/@/")) {
+            return p.permaid === current.prev.split("/@/")[1].trim();
+          }
+          return p.href === current?.prev;
+        }) || null;
     } else {
       previous = pageList[i - 1] || null;
     }