+++ packages/markdown/devBuild.mjs
@@ -6,10 +6,11 @@ const markdown = await fs.readFile("dev.md", "utf8");
 
 const result = await process(markdown, {
   root: "",
-  qrcode: true,
   config: {
     name: "Hyperbook Dokumenation",
     description: "Dokumentation für Hyperbook erstellt mit Hyperbook",
+    qrcode: true,
+    search: true,
     author: {
       name: "OpenPatch",
       url: "https://openpatch.org",
@@ -88,6 +89,7 @@ const result = await process(markdown, {
     },
     current: {
       name: "Markdown Referenz",
+      permaid: "md",
       index: 0,
       isEmpty: true,
       pages: [