+++ packages/fs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hyperbook/fs
 
+## 0.16.0
+
+### Minor Changes
+
+- [`27d2f47`](https://github.com/openpatch/hyperbook/commit/27d2f47e0e669b3738f62206bed767d29abce18a) Thanks [@mikebarkmin](https://github.com/mikebarkmin)! - Improve concat function to accept many strings
+
+## 0.15.0
+
+### Minor Changes
+
+- [`d9f0b71`](https://github.com/openpatch/hyperbook/commit/d9f0b711775195fc56d9706dc196edf577591b2e) Thanks [@mikebarkmin](https://github.com/mikebarkmin)! - Allow permaids to be set in the frontmatter of a page. The permaids can be use in links and in the frontmatter for next and prev. See the page configuration documentation for details. All pages with permaids are available at /@/[permaid].
+
 ## 0.14.1
 
 ### Patch Changes